recognition/requirements.txt
flask
numpy
opencv-python
python-dotenv
sqlalchemy
psycopg2
qreader